Fix superfluid + export utils methods in evm_utils
